+++ tests/test_mcmc.py
@@ -5,8 +5,6 @@ import pytest
 from numpy.linalg import LinAlgError
 from scipy.stats import norm
 from pycalphad import Database, variables as v
-from pycalphad.codegen.callables import build_callables
-from pycalphad.core.utils import instantiate_models
 
 from espei.error_functions.context import setup_context
 from espei.optimizers.opt_mcmc import EmceeOptimizer
@@ -34,12 +32,17 @@ def test_lnprob_calculates_multi_phase_probability_for_success(datasets_db):
     res = opt.predict([10], prior_rvs=[rv_zero()], symbols_to_fit=[param], zpf_kwargs=zpf_kwargs)
 
     assert np.isreal(res)
-    assert np.isclose(res, -31.309645520830344, rtol=1e-4)
+    assert not np.isinf(res)
+    assert np.isclose(res, -31.309645520830344, rtol=1e-6)
 
-    res_2 = opt.predict([10000000], prior_rvs=[rv_zero()], symbols_to_fit=[param], zpf_kwargs=zpf_kwargs)
+    # The purpose of this part is to test that the driving forces (and probability)
+    # are different than the case of VV0001 = 10.
+    res_2 = opt.predict([-10000000], prior_rvs=[rv_zero()], symbols_to_fit=[param], zpf_kwargs=zpf_kwargs)
 
-    assert not np.isclose(res_2, -31.309645520830344, rtol=1e-6)
-
+    assert np.isreal(res_2)
+    assert not np.isinf(res_2)
+    # Accept a large rtol becuase the results should be _very_ different
+    assert not np.isclose(res_2, -31.309645520830344, rtol=1e-2)
 
 def test_lnprob_calculates_single_phase_probability_for_success(datasets_db):
     """lnprob() succesfully calculates the probability from single phase data"""